requirements.txt
# Core Dependencies
pydantic>=2.0.0
pyyaml>=6.0

# Vector Database & RAG
chromadb>=0.4.0
sentence-transformers>=2.2.0

# API Framework
fastapi>=0.100.0
uvicorn>=0.23.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Development
pytest>=7.4.0
black>=23.0.0
